refactor(runner): separate kill command responsibilities
