+++ crates/runner/src/cmd/kill/orphan.rs
@@ -1,0 +1,1073 @@
+use std::path::PathBuf;
+use std::time::Duration;
+
+use tracing::info;
+
+use crate::process::{
+    self, DiscoveredProcesses, FirecrackerProcessIdentity, ProcessStat, ProcessStatRead,
+};
+
+use super::target::KillTarget;
+
+const ORPHAN_EXIT_TIMEOUT: Duration = Duration::from_secs(5);
+const ORPHAN_EXIT_POLL_INTERVAL: Duration = Duration::from_millis(50);
+
+#[derive(Debug)]
+pub(super) enum Outcome {
+    Killed(KillTarget),
+    AlreadyExited(KillTarget),
+    TerminationUnconfirmed {
+        target: KillTarget,
+        failure: OrphanExitFailure,
+    },
+    AlreadyExitedOrChanged(KillTarget),
+    SignalFailed(KillTarget),
+}
+
+pub(super) async fn terminate(target: &KillTarget) -> Outcome {
+    let identity = match validate_orphan_target(target).await {
+        OrphanTargetValidation::Valid { identity } => identity,
+        OrphanTargetValidation::AlreadyGone => {
+            return already_gone_orphan_outcome(target).await;
+        }
+        OrphanTargetValidation::Changed => {
+            return Outcome::AlreadyExitedOrChanged(target.clone());
+        }
+    };
+
+    match signal_process_group(target.pid, identity.pgid) {
+        ProcessGroupSignalResult::Signaled => match wait_for_orphan_exit(&identity).await {
+            Ok(()) => Outcome::Killed(target.clone()),
+            Err(failure) => Outcome::TerminationUnconfirmed {
+                target: target.clone(),
+                failure,
+            },
+        },
+        ProcessGroupSignalResult::AlreadyGone => already_gone_orphan_outcome(target).await,
+        ProcessGroupSignalResult::Failed => Outcome::SignalFailed(target.clone()),
+    }
+}
+
+pub(super) async fn confirmed_disappeared_outcome(
+    target: &KillTarget,
+    was_orphan: bool,
+) -> Option<Outcome> {
+    let discovered = process::discover_all_with_status().await;
+    if should_cleanup_disappeared_initial_orphan(
+        target,
+        was_orphan,
+        discovered.proc_scan_complete,
+        &discovered.processes,
+    ) && initial_process_confirmed_terminated(target).await
+    {
+        Some(Outcome::AlreadyExited(target.clone()))
+    } else {
+        None
+    }
+}
+
+async fn already_gone_orphan_outcome(target: &KillTarget) -> Outcome {
+    confirmed_disappeared_outcome(target, true)
+        .await
+        .unwrap_or_else(|| Outcome::AlreadyExitedOrChanged(target.clone()))
+}
+
+fn should_cleanup_disappeared_initial_orphan(
+    initial: &KillTarget,
+    was_orphan: bool,
+    proc_scan_complete: bool,
+    discovered_after_error: &DiscoveredProcesses,
+) -> bool {
+    proc_scan_complete
+        && was_orphan
+        && target_has_workspace_identity(initial)
+        && !discovered_has_same_or_unidentified_firecracker(initial, discovered_after_error)
+}
+
+fn target_has_workspace_identity(target: &KillTarget) -> bool {
+    match (&target.base_dir, &target.identity) {
+        (Some(base_dir), Some(identity)) => {
+            identity.sandbox_id == target.sandbox_id && identity.base_dir.as_ref() == Some(base_dir)
+        }
+        _ => false,
+    }
+}
+
+fn discovered_has_same_or_unidentified_firecracker(
+    initial: &KillTarget,
+    discovered: &DiscoveredProcesses,
+) -> bool {
+    discovered.firecrackers.iter().any(|process| {
+        process.sandbox_id == initial.sandbox_id || process.workspace_identity_incomplete()
+    })
+}
+
+enum OrphanTargetValidation {
+    Valid {
+        identity: FirecrackerProcessIdentity,
+    },
+    AlreadyGone,
+    Changed,
+}
+
+async fn validate_orphan_target(target: &KillTarget) -> OrphanTargetValidation {
+    let Some(identity) = &target.identity else {
+        tracing::warn!(
+            pid = target.pid,
+            sandbox_id = %target.sandbox_id,
+            "refusing orphan kill without process identity"
+        );
+        return OrphanTargetValidation::Changed;
+    };
+
+    if identity.pid != target.pid {
+        tracing::warn!(
+            pid = target.pid,
+            identity_pid = identity.pid,
+            "refusing orphan kill with inconsistent process identity"
+        );
+        return OrphanTargetValidation::Changed;
+    }
+
+    let stat = match process::read_process_stat_checked(target.pid).await {
+        ProcessStatRead::Found(stat) => stat,
+        ProcessStatRead::Missing => {
+            tracing::warn!(
+                pid = target.pid,
+                "orphan target disappeared before validation"
+            );
+            return OrphanTargetValidation::AlreadyGone;
+        }
+        ProcessStatRead::Unreadable(error) => {
+            tracing::warn!(
+                pid = target.pid,
+                %error,
+                "refusing orphan kill because process stat is unreadable"
+            );
+            return OrphanTargetValidation::Changed;
+        }
+        ProcessStatRead::Invalid => {
+            tracing::warn!(
+                pid = target.pid,
+                "refusing orphan kill because process stat is invalid"
+            );
+            return OrphanTargetValidation::Changed;
+        }
+    };
+    if !process_stat_matches_identity(identity, &stat) {
+        tracing::warn!(
+            pid = target.pid,
+            expected_pgid = identity.pgid,
+            current_pgid = stat.pgid,
+            expected_starttime = identity.starttime,
+            current_starttime = stat.starttime,
+            "refusing orphan kill after process identity changed"
+        );
+        return OrphanTargetValidation::Changed;
+    }
+    if !process::process_stat_is_live(&stat) {
+        tracing::warn!(
+            pid = target.pid,
+            state = %stat.state,
+            "orphan target already exited and is waiting to be reaped"
+        );
+        return OrphanTargetValidation::AlreadyGone;
+    }
+
+    let Some(cmdline) = process::read_cmdline(target.pid).await else {
+        tracing::warn!(
+            pid = target.pid,
+            "failed to read cmdline before orphan kill"
+        );
+        return classify_orphan_validation_after_unreadable_pid_fact(target.pid, identity).await;
+    };
+    if !process::is_firecracker_cmdline(&cmdline) {
+        tracing::warn!(
+            pid = target.pid,
+            "refusing orphan kill for non-firecracker cmdline"
+        );
+        return OrphanTargetValidation::Changed;
+    }
+
+    let cwd_info = process::read_cwd(target.pid)
+        .await
+        .and_then(|cwd| process::parse_workspace_cwd(&cwd));
+    if !orphan_identity_matches_facts(identity, &stat, true, cwd_info.as_ref()) {
+        tracing::warn!(
+            pid = target.pid,
+            sandbox_id = %target.sandbox_id,
+            "refusing orphan kill after workspace identity changed"
+        );
+        return classify_orphan_validation_after_unreadable_pid_fact(target.pid, identity).await;
+    }
+
+    let final_stat = match process::read_process_stat_checked(target.pid).await {
+        ProcessStatRead::Found(stat) => stat,
+        ProcessStatRead::Missing => {
+            tracing::warn!(
+                pid = target.pid,
+                "orphan target disappeared during validation"
+            );
+            return OrphanTargetValidation::AlreadyGone;
+        }
+        ProcessStatRead::Unreadable(error) => {
+            tracing::warn!(
+                pid = target.pid,
+                %error,
+                "refusing orphan kill because process stat became unreadable"
+            );
+            return OrphanTargetValidation::Changed;
+        }
+        ProcessStatRead::Invalid => {
+            tracing::warn!(
+                pid = target.pid,
+                "refusing orphan kill because process stat became invalid"
+            );
+            return OrphanTargetValidation::Changed;
+        }
+    };
+    if !process_stat_matches_identity(identity, &final_stat) {
+        tracing::warn!(
+            pid = target.pid,
+            expected_pgid = identity.pgid,
+            current_pgid = final_stat.pgid,
+            expected_starttime = identity.starttime,
+            current_starttime = final_stat.starttime,
+            "refusing orphan kill after process identity changed during validation"
+        );
+        return OrphanTargetValidation::Changed;
+    }
+    if !process::process_stat_is_live(&final_stat) {
+        tracing::warn!(
+            pid = target.pid,
+            state = %final_stat.state,
+            "orphan target exited during validation and is waiting to be reaped"
+        );
+        return OrphanTargetValidation::AlreadyGone;
+    }
+
+    OrphanTargetValidation::Valid {
+        identity: identity.clone(),
+    }
+}
+
+async fn classify_orphan_validation_after_unreadable_pid_fact(
+    pid: u32,
+    identity: &FirecrackerProcessIdentity,
+) -> OrphanTargetValidation {
+    match process::read_process_stat_checked(pid).await {
+        ProcessStatRead::Found(stat)
+            if process_stat_matches_identity(identity, &stat)
+                && !process::process_stat_is_live(&stat) =>
+        {
+            OrphanTargetValidation::AlreadyGone
+        }
+        ProcessStatRead::Found(stat) if process_stat_matches_identity(identity, &stat) => {
+            OrphanTargetValidation::Changed
+        }
+        ProcessStatRead::Found(_) => OrphanTargetValidation::Changed,
+        ProcessStatRead::Missing => OrphanTargetValidation::AlreadyGone,
+        ProcessStatRead::Unreadable(_) | ProcessStatRead::Invalid => {
+            OrphanTargetValidation::Changed
+        }
+    }
+}
+
+async fn initial_process_confirmed_terminated(target: &KillTarget) -> bool {
+    let Some(identity) = &target.identity else {
+        return false;
+    };
+    matches!(
+        classify_orphan_exit_observation(
+            identity,
+            process::read_process_stat_checked(target.pid).await,
+        ),
+        OrphanExitObservation::Terminated
+    )
+}
+
+#[derive(Debug, Eq, PartialEq)]
+enum OrphanExitObservation {
+    Live,
+    Terminated,
+    IdentityChanged {
+        expected_pgid: u32,
+        observed_pgid: u32,
+        expected_starttime: u64,
+        observed_starttime: u64,
+    },
+    Unverifiable(String),
+}
+
+#[derive(Debug, Eq, PartialEq)]
+pub(super) enum OrphanExitFailure {
+    TimedOut,
+    IdentityChanged {
+        expected_pgid: u32,
+        observed_pgid: u32,
+        expected_starttime: u64,
+        observed_starttime: u64,
+    },
+    Unverifiable(String),
+}
+
+impl std::fmt::Display for OrphanExitFailure {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::TimedOut => f.write_str("process is still live at the termination deadline"),
+            Self::IdentityChanged {
+                expected_pgid,
+                observed_pgid,
+                expected_starttime,
+                observed_starttime,
+            } => write!(
+                f,
+                "process identity changed while waiting for termination \
+                 (PGID {expected_pgid} -> {observed_pgid}, starttime \
+                 {expected_starttime} -> {observed_starttime})"
+            ),
+            Self::Unverifiable(error) => {
+                write!(f, "process termination could not be verified: {error}")
+            }
+        }
+    }
+}
+
+fn classify_orphan_exit_observation(
+    identity: &FirecrackerProcessIdentity,
+    read: ProcessStatRead,
+) -> OrphanExitObservation {
+    match read {
+        ProcessStatRead::Found(stat) if !process_stat_matches_identity(identity, &stat) => {
+            OrphanExitObservation::IdentityChanged {
+                expected_pgid: identity.pgid,
+                observed_pgid: stat.pgid,
+                expected_starttime: identity.starttime,
+                observed_starttime: stat.starttime,
+            }
+        }
+        ProcessStatRead::Found(stat) if process::process_stat_is_live(&stat) => {
+            OrphanExitObservation::Live
+        }
+        ProcessStatRead::Found(_) | ProcessStatRead::Missing => OrphanExitObservation::Terminated,
+        ProcessStatRead::Unreadable(error) => {
+            OrphanExitObservation::Unverifiable(error.to_string())
+        }
+        ProcessStatRead::Invalid => {
+            OrphanExitObservation::Unverifiable("process stat is invalid".into())
+        }
+    }
+}
+
+async fn wait_for_orphan_exit(
+    identity: &FirecrackerProcessIdentity,
+) -> Result<(), OrphanExitFailure> {
+    wait_for_orphan_exit_with(
+        identity,
+        ORPHAN_EXIT_TIMEOUT,
+        ORPHAN_EXIT_POLL_INTERVAL,
+        process::read_process_stat_checked,
+    )
+    .await
+}
+
+async fn wait_for_orphan_exit_with<ReadStat, ReadStatFuture>(
+    identity: &FirecrackerProcessIdentity,
+    timeout: Duration,
+    poll_interval: Duration,
+    mut read_stat: ReadStat,
+) -> Result<(), OrphanExitFailure>
+where
+    ReadStat: FnMut(u32) -> ReadStatFuture,
+    ReadStatFuture: std::future::Future<Output = ProcessStatRead>,
+{
+    let deadline = tokio::time::Instant::now() + timeout;
+    loop {
+        let failure =
+            match classify_orphan_exit_observation(identity, read_stat(identity.pid).await) {
+                OrphanExitObservation::Terminated => return Ok(()),
+                OrphanExitObservation::IdentityChanged {
+                    expected_pgid,
+                    observed_pgid,
+                    expected_starttime,
+                    observed_starttime,
+                } => {
+                    return Err(OrphanExitFailure::IdentityChanged {
+                        expected_pgid,
+                        observed_pgid,
+                        expected_starttime,
+                        observed_starttime,
+                    });
+                }
+                OrphanExitObservation::Live => OrphanExitFailure::TimedOut,
+                OrphanExitObservation::Unverifiable(error) => {
+                    OrphanExitFailure::Unverifiable(error)
+                }
+            };
+
+        let now = tokio::time::Instant::now();
+        if now >= deadline {
+            return Err(failure);
+        }
+        tokio::time::sleep(poll_interval.min(deadline - now)).await;
+    }
+}
+
+fn process_stat_matches_identity(
+    identity: &FirecrackerProcessIdentity,
+    stat: &ProcessStat,
+) -> bool {
+    stat.pgid == identity.pgid && stat.starttime == identity.starttime
+}
+
+fn orphan_identity_matches_facts(
+    identity: &FirecrackerProcessIdentity,
+    stat: &ProcessStat,
+    is_firecracker_cmdline: bool,
+    cwd_info: Option<&(String, PathBuf)>,
+) -> bool {
+    process_stat_matches_identity(identity, stat)
+        && is_firecracker_cmdline
+        && workspace_identity_matches(identity, cwd_info)
+}
+
+fn workspace_identity_matches(
+    identity: &FirecrackerProcessIdentity,
+    cwd_info: Option<&(String, PathBuf)>,
+) -> bool {
+    match (&identity.base_dir, cwd_info) {
+        (Some(expected_base_dir), Some((sandbox_id, base_dir))) => {
+            sandbox_id == &identity.sandbox_id && base_dir == expected_base_dir
+        }
+        _ => false,
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+enum ProcessGroupSignalResult {
+    Signaled,
+    AlreadyGone,
+    Failed,
+}
+
+/// Send `SIGKILL` to a validated process group.
+fn signal_process_group(pid: u32, pgid: u32) -> ProcessGroupSignalResult {
+    if pgid <= 1 {
+        tracing::warn!(pid, pgid, "refusing to signal system process group");
+        return ProcessGroupSignalResult::Failed;
+    }
+
+    let Ok(pgid_i32) = i32::try_from(pgid) else {
+        return ProcessGroupSignalResult::Failed;
+    };
+    if nix::unistd::getpgrp().as_raw() == pgid_i32 {
+        tracing::warn!(pid, pgid = pgid_i32, "refusing to signal own process group");
+        return ProcessGroupSignalResult::Failed;
+    }
+
+    match nix::sys::signal::killpg(
+        nix::unistd::Pid::from_raw(pgid_i32),
+        nix::sys::signal::Signal::SIGKILL,
+    ) {
+        Ok(()) => {
+            info!(pid, pgid = pgid_i32, "killed process group");
+            ProcessGroupSignalResult::Signaled
+        }
+        Err(nix::errno::Errno::ESRCH) => {
+            info!(
+                pid,
+                pgid = pgid_i32,
+                "process group already exited before signal"
+            );
+            ProcessGroupSignalResult::AlreadyGone
+        }
+        Err(e) => {
+            tracing::warn!(pid, pgid = pgid_i32, error = %e, "failed to kill process group");
+            ProcessGroupSignalResult::Failed
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::VecDeque;
+    use std::path::PathBuf;
+
+    use tokio::io::{AsyncBufReadExt, BufReader};
+
+    use super::super::test_support::{discovered_with_firecrackers, make_fc, make_target};
+    use super::*;
+    use crate::process::FirecrackerProcessInfo;
+    use crate::test_fixtures::ignored_child::{
+        ignored_child_test_env_guard_enabled, run_ignored_child_test,
+    };
+
+    const ORPHAN_KILL_CHILD_ENV: &str = "VM0_RUNNER_ORPHAN_KILL_TEST_CHILD";
+    const ORPHAN_KILL_READY_LINE: &str = "vm0 orphan kill test ready";
+
+    fn process_stat(identity: &FirecrackerProcessIdentity) -> ProcessStat {
+        process_stat_with_state(identity, 'S')
+    }
+
+    fn process_stat_with_state(identity: &FirecrackerProcessIdentity, state: char) -> ProcessStat {
+        ProcessStat {
+            state,
+            ppid: 7,
+            pgid: identity.pgid,
+            starttime: identity.starttime,
+        }
+    }
+
+    #[test]
+    fn disappeared_initial_orphan_with_identity_allows_cleanup() {
+        let initial = make_target(200, "sbox-123");
+        let discovered = discovered_with_firecrackers(vec![]);
+
+        assert!(should_cleanup_disappeared_initial_orphan(
+            &initial,
+            true,
+            true,
+            &discovered
+        ));
+    }
+
+    #[test]
+    fn incomplete_process_scan_rejects_disappeared_orphan_cleanup() {
+        let initial = make_target(200, "sbox-123");
+        let discovered = discovered_with_firecrackers(vec![]);
+
+        assert!(!should_cleanup_disappeared_initial_orphan(
+            &initial,
+            true,
+            false,
+            &discovered
+        ));
+    }
+
+    #[test]
+    fn disappeared_initial_managed_target_rejects_cleanup() {
+        let initial = make_target(200, "sbox-123");
+        let discovered = discovered_with_firecrackers(vec![]);
+
+        assert!(!should_cleanup_disappeared_initial_orphan(
+            &initial,
+            false,
+            true,
+            &discovered
+        ));
+    }
+
+    #[test]
+    fn disappeared_initial_without_workspace_identity_rejects_cleanup() {
+        let mut initial = make_target(200, "sbox-123");
+        initial.identity.as_mut().unwrap().base_dir = None;
+        let discovered = discovered_with_firecrackers(vec![]);
+
+        assert!(!should_cleanup_disappeared_initial_orphan(
+            &initial,
+            true,
+            true,
+            &discovered
+        ));
+    }
+
+    #[test]
+    fn disappeared_initial_with_same_sandbox_still_running_rejects_cleanup() {
+        let initial = make_target(200, "sbox-123");
+        let discovered = discovered_with_firecrackers(vec![make_fc(201, "sbox-123")]);
+
+        assert!(!should_cleanup_disappeared_initial_orphan(
+            &initial,
+            true,
+            true,
+            &discovered
+        ));
+    }
+
+    #[test]
+    fn disappeared_initial_with_unidentified_firecracker_rejects_cleanup() {
+        let initial = make_target(200, "sbox-123");
+        let discovered = discovered_with_firecrackers(vec![FirecrackerProcessInfo {
+            pid: 201,
+            ppid: None,
+            sandbox_id: "pid-201".into(),
+            base_dir: None,
+            identity: None,
+        }]);
+
+        assert!(!should_cleanup_disappeared_initial_orphan(
+            &initial,
+            true,
+            true,
+            &discovered
+        ));
+    }
+
+    #[test]
+    fn orphan_identity_facts_match() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let stat = process_stat(identity);
+        let cwd_info = ("sbox-123".to_string(), PathBuf::from("/data/r1"));
+
+        assert!(orphan_identity_matches_facts(
+            identity,
+            &stat,
+            true,
+            Some(&cwd_info)
+        ));
+    }
+
+    #[test]
+    fn orphan_identity_rejects_changed_starttime() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let stat = ProcessStat {
+            state: 'S',
+            ppid: 7,
+            pgid: identity.pgid,
+            starttime: identity.starttime + 1,
+        };
+        let cwd_info = ("sbox-123".to_string(), PathBuf::from("/data/r1"));
+
+        assert!(!orphan_identity_matches_facts(
+            identity,
+            &stat,
+            true,
+            Some(&cwd_info)
+        ));
+    }
+
+    #[test]
+    fn orphan_identity_rejects_changed_pgid() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let stat = ProcessStat {
+            state: 'S',
+            ppid: 7,
+            pgid: identity.pgid + 1,
+            starttime: identity.starttime,
+        };
+        let cwd_info = ("sbox-123".to_string(), PathBuf::from("/data/r1"));
+
+        assert!(!orphan_identity_matches_facts(
+            identity,
+            &stat,
+            true,
+            Some(&cwd_info)
+        ));
+    }
+
+    #[test]
+    fn process_stat_identity_match_requires_stable_pgid_and_starttime() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let matching = process_stat(identity);
+        let changed_starttime = ProcessStat {
+            state: 'S',
+            ppid: 7,
+            pgid: identity.pgid,
+            starttime: identity.starttime + 1,
+        };
+        let changed_pgid = ProcessStat {
+            state: 'S',
+            ppid: 7,
+            pgid: identity.pgid + 1,
+            starttime: identity.starttime,
+        };
+        let changed_ppid = ProcessStat {
+            state: 'S',
+            ppid: 9,
+            pgid: identity.pgid,
+            starttime: identity.starttime,
+        };
+
+        assert!(process_stat_matches_identity(identity, &matching));
+        assert!(process_stat_matches_identity(identity, &changed_ppid));
+        assert!(!process_stat_matches_identity(identity, &changed_starttime));
+        assert!(!process_stat_matches_identity(identity, &changed_pgid));
+    }
+
+    #[test]
+    fn orphan_identity_rejects_non_firecracker_cmdline() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let stat = process_stat(identity);
+        let cwd_info = ("sbox-123".to_string(), PathBuf::from("/data/r1"));
+
+        assert!(!orphan_identity_matches_facts(
+            identity,
+            &stat,
+            false,
+            Some(&cwd_info)
+        ));
+    }
+
+    #[test]
+    fn orphan_identity_rejects_changed_workspace() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let stat = process_stat(identity);
+        let cwd_info = ("sbox-other".to_string(), PathBuf::from("/data/r1"));
+
+        assert!(!orphan_identity_matches_facts(
+            identity,
+            &stat,
+            true,
+            Some(&cwd_info)
+        ));
+    }
+
+    #[test]
+    fn orphan_identity_rejects_missing_workspace_identity() {
+        let mut target = make_target(200, "sbox-123");
+        target.base_dir = None;
+        target.identity.as_mut().unwrap().base_dir = None;
+        let identity = target.identity.as_ref().unwrap();
+        let stat = process_stat(identity);
+
+        assert!(!orphan_identity_matches_facts(identity, &stat, true, None));
+    }
+
+    #[test]
+    fn zombie_process_stat_is_already_exited() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let zombie = ProcessStat {
+            state: 'Z',
+            ppid: 7,
+            pgid: identity.pgid,
+            starttime: identity.starttime,
+        };
+
+        assert!(process_stat_matches_identity(identity, &zombie));
+        assert!(!process::process_stat_is_live(&zombie));
+    }
+
+    #[test]
+    fn dead_process_stat_is_already_exited() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let dead = ProcessStat {
+            state: 'X',
+            ppid: 7,
+            pgid: identity.pgid,
+            starttime: identity.starttime,
+        };
+
+        assert!(process_stat_matches_identity(identity, &dead));
+        assert!(!process::process_stat_is_live(&dead));
+    }
+
+    #[test]
+    fn orphan_exit_observation_distinguishes_live_and_terminal_states() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+
+        assert_eq!(
+            classify_orphan_exit_observation(
+                identity,
+                ProcessStatRead::Found(process_stat(identity)),
+            ),
+            OrphanExitObservation::Live
+        );
+        for state in ['Z', 'X', 'x'] {
+            assert_eq!(
+                classify_orphan_exit_observation(
+                    identity,
+                    ProcessStatRead::Found(process_stat_with_state(identity, state)),
+                ),
+                OrphanExitObservation::Terminated
+            );
+        }
+        assert_eq!(
+            classify_orphan_exit_observation(identity, ProcessStatRead::Missing),
+            OrphanExitObservation::Terminated
+        );
+    }
+
+    #[test]
+    fn orphan_exit_observation_rejects_identity_changes() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let changed = ProcessStat {
+            pgid: identity.pgid + 1,
+            starttime: identity.starttime + 1,
+            ..process_stat(identity)
+        };
+
+        assert_eq!(
+            classify_orphan_exit_observation(identity, ProcessStatRead::Found(changed)),
+            OrphanExitObservation::IdentityChanged {
+                expected_pgid: identity.pgid,
+                observed_pgid: identity.pgid + 1,
+                expected_starttime: identity.starttime,
+                observed_starttime: identity.starttime + 1,
+            }
+        );
+    }
+
+    #[test]
+    fn orphan_exit_observation_rejects_unverifiable_reads() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+
+        assert!(matches!(
+            classify_orphan_exit_observation(
+                identity,
+                ProcessStatRead::Unreadable(std::io::Error::from(
+                    std::io::ErrorKind::PermissionDenied,
+                )),
+            ),
+            OrphanExitObservation::Unverifiable(_)
+        ));
+        assert_eq!(
+            classify_orphan_exit_observation(identity, ProcessStatRead::Invalid),
+            OrphanExitObservation::Unverifiable("process stat is invalid".into())
+        );
+    }
+
+    #[tokio::test]
+    async fn orphan_exit_wait_observes_delayed_termination() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let mut reads = VecDeque::from([
+            ProcessStatRead::Found(process_stat(identity)),
+            ProcessStatRead::Found(process_stat_with_state(identity, 'Z')),
+        ]);
+
+        let result = wait_for_orphan_exit_with(
+            identity,
+            Duration::from_secs(1),
+            Duration::ZERO,
+            move |_| std::future::ready(reads.pop_front().unwrap()),
+        )
+        .await;
+
+        assert_eq!(result, Ok(()));
+    }
+
+    #[tokio::test]
+    async fn orphan_exit_wait_times_out_while_identity_is_live() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+
+        let result = wait_for_orphan_exit_with(identity, Duration::ZERO, Duration::ZERO, |_| {
+            std::future::ready(ProcessStatRead::Found(process_stat(identity)))
+        })
+        .await;
+
+        assert_eq!(result, Err(OrphanExitFailure::TimedOut));
+    }
+
+    #[tokio::test]
+    async fn orphan_exit_wait_fails_when_identity_changes() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let changed = ProcessStat {
+            pgid: identity.pgid + 1,
+            starttime: identity.starttime + 1,
+            ..process_stat(identity)
+        };
+
+        let result =
+            wait_for_orphan_exit_with(identity, Duration::from_secs(1), Duration::ZERO, |_| {
+                std::future::ready(ProcessStatRead::Found(changed.clone()))
+            })
+            .await;
+
+        let Err(failure) = result else {
+            panic!("changed process identity should fail the exit wait");
+        };
+        assert_eq!(
+            failure,
+            OrphanExitFailure::IdentityChanged {
+                expected_pgid: identity.pgid,
+                observed_pgid: identity.pgid + 1,
+                expected_starttime: identity.starttime,
+                observed_starttime: identity.starttime + 1,
+            }
+        );
+        assert_eq!(
+            failure.to_string(),
+            "process identity changed while waiting for termination \
+             (PGID 1200 -> 1201, starttime 123456 -> 123457)"
+        );
+    }
+
+    #[tokio::test]
+    async fn orphan_exit_wait_recovers_from_transient_unverifiable_read() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+        let mut reads = VecDeque::from([ProcessStatRead::Invalid, ProcessStatRead::Missing]);
+
+        let result = wait_for_orphan_exit_with(
+            identity,
+            Duration::from_secs(1),
+            Duration::ZERO,
+            move |_| std::future::ready(reads.pop_front().unwrap()),
+        )
+        .await;
+
+        assert_eq!(result, Ok(()));
+    }
+
+    #[tokio::test]
+    async fn orphan_exit_wait_fails_when_observation_remains_unverifiable() {
+        let target = make_target(200, "sbox-123");
+        let identity = target.identity.as_ref().unwrap();
+
+        let result = wait_for_orphan_exit_with(identity, Duration::ZERO, Duration::ZERO, |_| {
+            std::future::ready(ProcessStatRead::Invalid)
+        })
+        .await;
+
+        let Err(failure) = result else {
+            panic!("unverifiable process state should fail at the exit deadline");
+        };
+        assert_eq!(
+            failure,
+            OrphanExitFailure::Unverifiable("process stat is invalid".into())
+        );
+        assert_eq!(
+            failure.to_string(),
+            "process termination could not be verified: process stat is invalid"
+        );
+    }
+
+    #[tokio::test]
+    async fn orphan_kill_validates_signals_and_waits_for_real_exit() {
+        run_ignored_child_test(
+            "cmd::kill::orphan::tests::orphan_kill_validates_signals_and_waits_for_real_exit_child",
+            (ORPHAN_KILL_CHILD_ENV, "1"),
+            &[],
+            Duration::from_secs(30),
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    #[ignore = "spawned by orphan_kill_validates_signals_and_waits_for_real_exit"]
+    async fn orphan_kill_validates_signals_and_waits_for_real_exit_child() {
+        if !ignored_child_test_env_guard_enabled((ORPHAN_KILL_CHILD_ENV, "1")) {
+            return;
+        }
+
+        let base_dir = tempfile::tempdir().unwrap();
+        let sandbox_id = "test-sandbox";
+        let workspace = base_dir.path().join("workspaces").join(sandbox_id);
+        tokio::fs::create_dir_all(&workspace).await.unwrap();
+        let firecracker = base_dir.path().join("firecracker");
+        std::os::unix::fs::symlink("/bin/sh", &firecracker).unwrap();
+
+        let mut child = tokio::process::Command::new(&firecracker)
+            .arg("-c")
+            .arg("printf '%s\\n' \"$1\"; IFS= read -r _")
+            .arg("vm0-orphan-kill-test")
+            .arg(ORPHAN_KILL_READY_LINE)
+            .current_dir(&workspace)
+            .process_group(0)
+            .kill_on_drop(true)
+            .stdin(std::process::Stdio::piped())
+            .stdout(std::process::Stdio::piped())
+            .stderr(std::process::Stdio::null())
+            .spawn()
+            .unwrap();
+        let stdout = child.stdout.take().unwrap();
+        let mut stdout_lines = BufReader::new(stdout).lines();
+        let ready_line = tokio::time::timeout(Duration::from_secs(5), stdout_lines.next_line())
+            .await
+            .expect("fake firecracker readiness timed out")
+            .expect("read fake firecracker readiness")
+            .expect("fake firecracker exited before readiness");
+        assert_eq!(ready_line, ORPHAN_KILL_READY_LINE);
+
+        let pid = child.id().unwrap();
+        let ProcessStatRead::Found(stat) = process::read_process_stat_checked(pid).await else {
+            panic!("spawned process stat should be readable");
+        };
+        let target = KillTarget {
+            pid,
+            ppid: None,
+            run_id: None,
+            sandbox_id: sandbox_id.into(),
+            base_dir: Some(base_dir.path().to_path_buf()),
+            identity: Some(FirecrackerProcessIdentity {
+                pid,
+                pgid: stat.pgid,
+                starttime: stat.starttime,
+                sandbox_id: sandbox_id.into(),
+                base_dir: Some(base_dir.path().to_path_buf()),
+            }),
+        };
+
+        let outcome = terminate(&target).await;
+        assert!(
+            matches!(&outcome, Outcome::Killed(killed) if killed == &target),
+            "unexpected orphan kill outcome: {outcome:?}"
+        );
+
+        let status = child.wait().await.unwrap();
+        assert!(!status.success());
+    }
+
+    #[test]
+    fn signal_process_group_rejects_zero_pgid() {
+        assert_eq!(
+            signal_process_group(1234, 0),
+            ProcessGroupSignalResult::Failed
+        );
+    }
+
+    #[test]
+    fn signal_process_group_rejects_init_pgid() {
+        assert_eq!(
+            signal_process_group(1234, 1),
+            ProcessGroupSignalResult::Failed
+        );
+    }
+
+    #[test]
+    fn signal_process_group_rejects_own_pgid() {
+        let current_pgid = u32::try_from(nix::unistd::getpgrp().as_raw()).unwrap();
+
+        assert_eq!(
+            signal_process_group(1234, current_pgid),
+            ProcessGroupSignalResult::Failed
+        );
+    }
+
+    #[test]
+    fn signal_process_group_reports_already_gone_for_missing_group() {
+        let missing_pgid = i32::MAX as u32;
+
+        assert_eq!(
+            signal_process_group(1234, missing_pgid),
+            ProcessGroupSignalResult::AlreadyGone
+        );
+    }
+
+    #[tokio::test]
+    async fn orphan_kill_nonexistent_pid_reports_gone_when_cleanup_is_safe() {
+        // u32::MAX exceeds any valid PID — /proc/{pid}/stat won't exist
+        let target = KillTarget {
+            pid: u32::MAX,
+            ppid: None,
+            run_id: None,
+            sandbox_id: "sbox-missing".into(),
+            base_dir: Some(PathBuf::from("/data/r1")),
+            identity: Some(FirecrackerProcessIdentity {
+                pid: u32::MAX,
+                pgid: 1234,
+                starttime: 123456,
+                sandbox_id: "sbox-missing".into(),
+                base_dir: Some(PathBuf::from("/data/r1")),
+            }),
+        };
+
+        assert!(matches!(
+            terminate(&target).await,
+            Outcome::AlreadyExited(_)
+        ));
+    }
+}
